requirements.txt
streamlit
plotly
pandas
numpy
scipy
Pillow
matplotlib
reportlab
streamlit-lottie
requests
opencv-python-headless
cellpose
torch
torchvision
scikit-image